Remove custom start applications, provide one DemoTournamentApplication which will start a number of players with different strategies, each on there own thread.
