+++ build.gradle
@@ -1,7 +1,7 @@
 apply plugin: 'java'
 apply plugin: 'application'
 
-mainClassName = "com.zuehlke.jasschallenge.RandomBot"
+mainClassName = "com.zuehlke.jasschallenge.DemoTournamentApplication"
 
 compileJava {
   sourceCompatibility = '1.8'
