+++ src/main/java/com/zuehlke/jasschallenge/client/RemoteGame.java
@@ -9,10 +9,12 @@ import org.eclipse.jetty.websocket.client.WebSocketClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
 import java.net.URI;
+import java.net.URISyntaxException;
 import java.util.concurrent.TimeUnit;
 
-public class RemoteGame implements Game {
+public class RemoteGame implements Game, Runnable {
 
     private static final Logger logger = LoggerFactory.getLogger(RemoteGame.class);
     private static final int CLOSE_TIMEOUT_MIN = 30;
@@ -29,7 +31,12 @@ public class RemoteGame implements Game {
     }
 
     @Override
-    public void start() throws Exception {
+    public void start() {
+        new Thread(this).start();
+    }
+
+    @Override
+    public void run() {
         final WebSocketClient client = new WebSocketClient();
         try {
             RemoteGameSocket socket = new RemoteGameSocket(new GameHandler(player, sessionType, sessionName));
@@ -40,9 +47,14 @@ public class RemoteGame implements Game {
             client.connect(socket, uri, request);
             logger.debug("Connecting to: {}", uri);
             socket.awaitClose(CLOSE_TIMEOUT_MIN, TimeUnit.MINUTES);
+        } catch (Exception e) {
+            e.printStackTrace();
         } finally {
-            client.stop();
+            try {
+                client.stop();
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
         }
     }
-
 }
